+++ player/src/lib.rs
@@ -237,7 +237,7 @@ impl GlobalPlay for wgc::global::Global {
                     let module = ron::de::from_str(&code).unwrap();
                     wgc::pipeline::ShaderModuleSource::Naga(module)
                 } else {
-                    panic!("Unknown shader {}", data);
+                    panic!("Unknown shader {data}");
                 };
                 let (_, error) = self.device_create_shader_module(device, &desc, source, Some(id));
                 if let Some(e) = error {